==> area_dictionary/Старая_квартирография/step_2_replacement.py <==
import json
from collections import defaultdict

# Словарь замен ключей
replacements = {
    'Внуково парк': 'Внуково Парк',
    'Алтуфьевcкое 53': 'Алтуфьевское 53',
    'Артель': 'ARTEL',
    'Восемь кленов': '8 Кленов',
    'Артисан': 'Artisan',
    'Armani Casa': 'Armani/Casa Moscow Residences',
    'Адрес (Т-отель)': 'Adres',
    'Cult': 'CULT',
    'Forum': 'FORUM',
    'АФИ парк Воронцовский': 'AFI Park Воронцовский',
    'Афи Тауэр': 'Afi Tower',
    'Баланс': 'balance',
    'Барклая,6': 'Барклая 6',
    'По ул. Макаренко': 'по ул. Макаренко',
    'Чкалов': 'Дом Chkalov',
    'Метроном': 'Квартал Метроном',
    'Квартал Румянцево (Мелисад)': 'Квартал Румянцево',
    'Квартал на воде (Графика на воде)': 'Квартал на воде',
    'Космо 4-22': 'Космо 4/22',
    'Кутузовград 2': 'Кутузов Град',
    'Аквилон Митино': 'Митино',
    'Найс Лофт': 'Найс лофт',
    'Мытищи парк': 'Мытищи Парк',
    'Новоград павлино': 'Новоград Павлино',
    'Новое замитино': 'Новое Замитино',
    'Облака 2.0': 'Облака',
    'Москворецкий': 'Москворецкий (Тучково)',
    'РГ. Октябрьское поле': 'Октябрьское поле',
    'Петровский парк 2': 'Петровский парк',
    'Проспект 39 (Дистрикт 39)': 'Проспект 39',
    'Тропарево парк': 'Тропарево Парк',
    'Прибрежный парк': 'Прибрежный Парк',
    'Филатов Луг': 'Филатов луг',
    'Пречистенка, 8': 'Пречистенка 8',
    'Цветочные поляны_new': 'Цветочные Поляны',
    'ЭвоПарк Сокольники': 'Evopark Сокольники',
    'Амбер Сити': 'Amber City',
    'Аквилон Сигнал': 'SIGNAL',
    'Эво': 'ЭВО',
    'Аквилон Бисайд': 'BESIDE',
    'Апсайд Тауэрс': 'Upside Towers',
    'Учебный 9': 'в пер. Учебный',
    'Ситизен': 'CITYZEN',
    'ULTIMA': 'ultima',
    'Патрикс': 'The Patricks',
    'Спутник': 'Sky Sputnik',
    'Франк хаус': "Frank's house",
    'Сити Бэй': 'City Bay',
    'Долгоруковская': 'Долгоруковская 25',
    'Горки парк_2023': 'Горки Парк',
    'Эхэд': 'AHEAD',
    'Кристал крост': 'CRYSTAL Трилогия',
    'Трилогия CRYSTAL': 'CRYSTAL Трилогия',
    'Культ': 'CULT',
    'Кэпитал Тауэрс': 'Capital Towers',
    'Шампань': 'Champine',
    'Клементин': 'Clementine',
    'Дрим Рива_нет цен': 'Dream Riva',
    'Дрим Тауэрс': 'Dream Towers',
    'Эра': 'ERA',
    'МИRА': 'МИRA',
    'Крылатская 33': 'Клубный квартал Крылатская 33',
    'Citi-Mix Новокосино': 'СитиМикс Новокосино',
    'Ново Молоково': 'Ново-Молоково',
    'Седьмое небо': '7 небо',
    'Олимп': 'Олимп (Хотьково)',
    'Элевен': 'Eleven',
    'Самоцветы Мира': 'Самоцветы мира',
    'Мой город (Электрогорск)_продан': 'Мой город',
    'Одинград Семейный': 'Одинград. Квартал семейный',
    'Эмоушн': 'Emotion',
    'Энитео': 'Eniteo',
    'Эвер': 'Ever',
    'ЭвоПарк Измайлово': 'Evopark Измайлово',
    'Форум': 'FORUM',
    'Майданово Парк': 'Майданово парк',
    'Фэймос': 'Famous',
    'Сколково Ван (Заречный Квартал)': 'Skolkovo one',
    'Форевилль': 'Foreville',
    'Форивер': 'Foriver',
    'Форивер Резиденс': 'Foriver Residence',
    'Красная горка': 'Красная Горка',
    'Фулл Хаус': 'Full House',
    'Хедлайнер': 'headliner',
    'Хайд': 'Hide',
    'Никольский квартал Отрада': 'Никольский квартал',
    'Хай Лайф': 'High Life',
    'Холланд парк': 'Holland park',
    'Инсайдер': 'INSIDER',
    'Инди Тауэрз': 'Indy Towers',
    'Инджой': 'Injoy',
    'Джойс': 'Jois',
    'Кинг энд Санс': 'KING & SONS',
    'Влюберцы апарт-комплекс': 'Влюберцы',
    'Долина яузы': 'Долина Яузы',
    'Кинетик': 'Kinetik',
    'Бакеево парк': 'Бакеево Парк',
    'ЗИЛАРТ': 'Зиларт',
    'Заречье парк': 'Заречье Парк',
    'Белая дача парк': 'Белая Дача парк',
    'Белый град': 'Белый Град',
    'Богдановский лес': 'Богдановский Лес',
    'Бутово парк 2Б': 'Бутово парк 2',
    'Левел Бауманская': 'Level Бауманская',
    'Каштановая роща': 'Каштановая Роща',
    'Левел Звенигородская': 'Level Звенигородская',
    'Левел Мичуринский': 'Level Мичуринский',
    'Левел Нагатинская': 'Level Нагатинская',
    'Левел Нижегородская': 'Level Нижегородская',
    'Левел Причальный': 'Level Причальный',
    'Левел Селигерская': 'Level Селигерская',
    'Левел Стрешнево': 'Level Стрешнево',
    'Левел Южнопортовая': 'Level Южнопортовая',
    'Левел Южнопортовая_1': 'Level Южнопортовая',
    'Левенсон': 'Levenson',
    'Лайф тайм (Онест)': 'Life Time',
    'Лайф Варшавская': 'Life Варшавская',
    'Лаки 1905': 'Lucky',
    'Лаки': 'Lucky',
    'КИТ': 'Кит',
    'Лужники Коллекшн': 'Luzhniki Collection',
    'Май Приорити Басманная': 'MYPRIORITY Басманный',
    'Май Приорити Дубровка': 'MYPRIORITY Дубровка',
    'Май Приорити Павелецкая': 'MYPRIORITY Павелецкая',
    'Мэйнстрит': 'Mainstreet',
    'МОД': 'Mod',
    'Космический квартал': 'Космический Квартал',
    'Моментс': 'Moments',
    'Котельники парк': 'Котельники Парк',
    'Нева Тауэрс': 'NEVA TOWERS',
    'Нова': 'NOVA',
    'Садовническая,69': 'Садовническая 69',
    'Нагатино айленд': 'Nagatino i-Land',
    'Наметкин Тауэр': 'Nametkin tower',
    'Оникс': 'Onyx',
    'Опус': 'Opus',
    'Phantom': 'PHANTOM',
    'Портленд': 'Portland',
    'Прайд': 'Pride',
    'Ред 7': 'RED7',
    'Ракурс': 'Rakurs',
    'Репаблик': 'Republic',
    'Ривер парк Тауэрс Кутузовский': 'River Park Towers Кутузовский',
    'Риверскай': 'RiverSky',
    'Ривер парк': 'Riverpark',
    'Ривер Парк Королев': 'Ривер парк',
    'Ривер Хаус': 'Ривер хаус',
    'Роттердам': 'Rotterdam',
    'Дмитровка 9': 'Большая Дмитровка IX',
    'Две столицы': 'Две Столицы',
    'Деко Резиденс': 'Deco Residence',
    'Рублевский квартал': 'Рублевский Квартал',
    'Рашн Дизайн Дистрикт': 'Russian Design District',
    'SVET': 'Свет',
    'Сердце столицы': 'Сердце Столицы',
    'Сезар Сити': 'SEZAR CITY',
    'Sole Hill': 'SOLE HILL',
    'Селф': 'Self',
    'Академик (Троицк)': 'Академик',
    'Шагал': 'Shagal',
    'Шифт': 'Shift',
    'Веспер Тверская': 'Vesper Tverskaya',
    'Олл ин Салют': 'Salut',
    'Диалог': 'Dialog',
    'Никитский,6': 'Никитский, 6',
    'Логос': 'Logos',
    'Союз': 'СОЮЗ',
    'Тессинский,1': 'Тессинский 1',
    'Скай Гарден': 'Sky Garden',
    'Скай Хаус': 'Sky House',
    'Скай Вью': 'SkyView',
    'Соколин Парк': 'Sokolinn Park',
    'Солнечный Парк': 'Солнечный Парк 2',
    'Соул': 'Soul',
    'Спрингс': 'Springs',
    'Стеллар Сити': 'Stellar City',
    'Сидней сити': 'Sydney City',
    'Сидней Прайм': 'Sydney Prime',
    'Симфони 34': 'Symphony 34',
    'Тэйт': 'Tate',
    'Лэйк': 'The LAKE',
    'Топхиллс': 'TopHills',
    'Турандот': 'Turandot Residences',
    'Твелв': 'Twelve',
    'Уно. Головинские пруды': 'UNO. Головинские пруды',
    'Уно. Соколиная Гора': 'UNO. Соколиная гора',
    'Веер': 'VEER',
    'Веспер Кутузовский': 'Vesper Кутузовский',
    'Виктори Парк Резиденс': 'Victory Park Residences',
    'Войс Тауэрс': 'Voice Towers',
    'Ice Towers': 'ICE TOWERS',
    'Воксхолл': 'Voxhall',
    'Левел Лесной': 'Level Лесной',
    'Вейв': 'WAVE',
    'Вау': 'WOW',
    'Вест Гарден_2020': 'West Garden',
    'Вест Тауэр': 'West Tower',
    'Вудс': 'Woods',
    'Вэлби': 'Wellbe',
    'Ай лав - Счастье вместе': 'ilove',
    'Монодом Лэйк': 'Апарт-комплекс Monodom Lake',
    'Дом 22': 'ДОМ XXII',
    'Дом XXII': 'ДОМ XXII',
    'Движение.Говорово': 'Движение. Говорово',
    '1-ый Ленинградский': '1-й Ленинградский',
    'Иртышский квартал (2-й Иртышский)': '2-й Иртышский',
    'Дом Малевич': 'Дом MALEVICH',
    'Звезды Арбата (Новый Арбат 32)': 'Звёзды Арбата',
    'Ильинка 3_8': 'Ильинка 3/8',
    'Русич_Кантемировский': 'Кантемировский',
    'Квартал Монс': 'Квартал МОНС',
    'Кронштадсткий 9': 'Кронштадтский 9',
    'Лаврушинский 5': 'Лаврушинский',
    'Эко Видное 2.0': 'У реки. Эко Видное 2.0',
    'Новая Звезда-2': 'Новая Звезда',
    'Обручева, 30': 'Обручева 30',
    'Октябрьская 98 (Миррорздание)': 'Октябрьская 98',
    'Парк легенд': 'Парк Легенд',
    'Подольские кварталы': 'Подольские Кварталы',
    'Пятницкое, 58': 'Пятницкое, 58',
    'Резиденции замоскворечье': 'Резиденции Замоскворечье',
    'Русские сезоны': 'Русские Сезоны',
    'СберСити в Рублево-Архангельском': 'Сберсити',
    'Северный порт. Береговые кварталы.': 'Северный порт. Береговые кварталы',
    'Страна Заречная': 'Страна. Заречная',
    'Страна Озерная': 'Страна. Озерная',
    'Стремянный 2 (Амарант)': 'Стремянный 2',
    'Тессинский 1': 'Тессинский, 1',
    'РЕКА': 'Река',
    'Обыденский 1': 'Обыденский',
    'Юту': 'U2',
    'Pave': 'PAVE',
    'Истомкино парк': 'Истомкино Парк',
    'Апрелевка клаб': 'Апрелевка Клаб',
    'Высокий Берег': 'Высокий берег',
    'Новая Алексеевская Роща': 'Новая Алексеевская роща',
    'Гармония Парк': 'Гармония парк'

}


with open('typology.json', 'r', encoding='utf-8') as f:
    data = json.load(f)

updated_data = {}
replacements_count = 0
lowered_count = 0
conflicts = []

for old_key, inner_dict in data.items():
    # сначала применяем replacements
    new_key = replacements.get(old_key, old_key)

    # сохраняем до/после lower
    before_lower = new_key
    new_key = new_key.lower()

    if before_lower != new_key:
        lowered_count += 1

    if new_key != old_key:
        replacements_count += 1

    # Проверка: ключ уже есть
    if new_key in updated_data:
        for area, value in inner_dict.items():
            if area in updated_data[new_key]:
                conflicts.append((new_key, area))
            updated_data[new_key][area] = value
    else:
        updated_data[new_key] = inner_dict.copy()

# Сохраняем
with open('typology_all_updated.json', 'w', encoding='utf-8') as f:
    json.dump(updated_data, f, ensure_ascii=False, indent=2)

print(f"🔁 Всего изменено ключей (replacements + lower): {replacements_count}")
print(f"🔡 Только из-за lower изменилось: {lowered_count}")
print(f"⚠️ Обнаружено конфликтов: {len(conflicts)}")
